pyproject.toml
[project]
name = "agent-voice-bridge"
version = "0.1.0"
description = "Real-time voice calls with AI using Twilio + Gemini/OpenAI"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Evan Senter", email = "evansenter@gmail.com" }]
keywords = ["voice", "ai", "twilio", "gemini", "realtime"]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "google-genai>=1.0.0",
    "twilio>=9.10.0",
    "python-multipart>=0.0.22",
    "audioop-lts>=0.2.0",  # audioop backport for Python 3.13+
]

[project.optional-dependencies]
openai = ["openai>=1.50.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
]

[project.scripts]
voice-bridge = "agent_voice_bridge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_voice_bridge"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
